+++ storage/fuzz/fuzz_targets/archive_operations.rs
@@ -4,7 +4,7 @@ use arbitrary::Arbitrary;
 use commonware_runtime::{deterministic, Runner};
 use commonware_storage::{
     archive::{Archive, Config, Identifier},
-    index::translator::EightCap,
+    translator::EightCap,
 };
 use commonware_utils::array::FixedBytes;
 use libfuzzer_sys::fuzz_target;
