+++ storage/src/archive/mod.rs
@@ -114,7 +114,7 @@
 //! use commonware_runtime::{Spawner, Runner, deterministic};
 //! use commonware_cryptography::hash;
 //! use commonware_storage::{
-//!     index::translator::FourCap,
+//!     translator::FourCap,
 //!     archive::{Archive, Config},
 //! };
 //!
@@ -142,7 +142,7 @@
 //! ```
 
 mod storage;
-pub use crate::index::Translator;
+pub use crate::translator::Translator;
 pub use storage::{Archive, Identifier};
 use thiserror::Error;
 
@@ -198,8 +198,8 @@ pub struct Config<T: Translator, C> {
 mod tests {
     use super::*;
     use crate::{
-        index::translator::{FourCap, TwoCap},
         journal::Error as JournalError,
+        translator::{FourCap, TwoCap},
     };
     use commonware_codec::{varint::UInt, DecodeExt, EncodeSize, Error as CodecError};
     use commonware_macros::test_traced;
