+++ storage/src/index/mod.rs
@@ -15,31 +15,11 @@
 
 mod storage;
 pub use storage::{Cursor, Index};
-pub mod translator;
-
-use std::hash::{BuildHasher, Hash};
-
-/// Translate keys into an internal representation used by `Index`.
-///
-/// # Warning
-///
-/// The output of `transform` is used as the key in a hash table. If the output is not uniformly
-/// distributed, the performance of [Index] will degrade substantially.
-pub trait Translator: Clone + BuildHasher {
-    /// The type of the internal representation of keys.
-    ///
-    /// Although `Translator` is a [BuildHasher], the `Key` type must still implement [Hash] for compatibility
-    /// with the [std::collections::HashMap] used internally by [Index].
-    type Key: Eq + Hash + Copy;
-
-    /// Transform a key into its internal representation.
-    fn transform(&self, key: &[u8]) -> Self::Key;
-}
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::index::translator::TwoCap;
+    use crate::translator::TwoCap;
     use commonware_macros::test_traced;
     use commonware_runtime::{deterministic, Metrics};
     use rand::Rng;
