+++ storage/src/lib.rs
@@ -14,3 +14,4 @@ pub mod metadata;
 pub mod mmr;
 pub mod ordinal;
 pub mod rmap;
+pub mod translator;
