+++ storage/src/translator.rs
@@ -1,21 +1,37 @@
 //! Primitive implementations of [Translator].
 
-use super::Translator;
-use std::hash::{BuildHasher, Hasher};
+use std::hash::{BuildHasher, Hash, Hasher};
+
+/// Translate keys into a new representation (often a smaller one).
+///
+/// # Warning
+///
+/// The output of [Translator::transform] is often used as a key in a hash table. If the output is not
+/// uniformly distributed, the performance of said hash table will degrade substantially.
+pub trait Translator: Clone + BuildHasher {
+    /// The type of the internal representation of keys.
+    ///
+    /// Although [Translator] is a [BuildHasher], the `Key` type must still implement [Hash] for compatibility
+    /// with any hash table that wraps [Translator].
+    type Key: Eq + Hash + Copy;
+
+    /// Transform a key into its new representation.
+    fn transform(&self, key: &[u8]) -> Self::Key;
+}
 
 /// A “do-nothing” hasher for `uint`.
 ///
-/// [super::Index] typically stores keys that are **already hashed** (shortened by the [Translator]).
-/// Re-hashing them with SipHash (by [std::collections::HashMap]) would waste CPU, so we give `HashMap`
-/// this identity hasher instead:
+/// Most users typically store keys that are **already hashed** (shortened by the [Translator]).
+/// Re-hashing them with SipHash (by [std::collections::HashMap]) would waste CPU, so we give
+/// [std::collections::HashMap] this identity hasher instead:
 ///
-/// * `write_u8`, `write_u16`, `write_u32`, `write_u64` copies the input into an internal field;
-/// * `finish()` returns that value unchanged.
+/// * [Hasher::write_u8], [Hasher::write_u16], [Hasher::write_u32], [Hasher::write_u64] copies the input into an internal field;
+/// * [Hasher::finish] returns that value unchanged.
 ///
 /// # Warning
 ///
 /// This hasher is not suitable for general use. If the hasher is called over some type that is not
-/// `u8`, `u16`, `u32` or `u64`, it will panic.
+/// [u8], [u16], [u32] or [u64], it will panic.
 #[derive(Default, Clone)]
 pub struct UintIdentity {
     value: u64,
